chore: remove docker
For some researches it does not seem interesting to use the docker in this case.

I had several problems of privilege when trying to mount partitions like tmpfs and also to run an init system like systemd inside the container.

An alternative would be to use VM instead of containers.