feat(airootfs): hook for vendor exclusion after compilation
When the UnaveOS source is synchronized, links are created at their respective destinations. The problem is that as our vendor is separated after compilation, the links are broken.

To get around this the vendor is now a core module as the construction script can only resolve directories within the airootf directory.

As long as we don't have a better solution, the hook will be executed.

Reference of the code snippet mentioned:
https://github.com/archlinux/archiso/blob/master/archiso/mkarchiso#L318-L328